Add /yarn.lock to .gitignore
A new entry, /yarn.lock, was added to the .gitignore file. This is done to prevent the yarn lock file from being tracked by Git as it contains specific versions of installed packages, and may cause inconsistency with other developers' environments if pushed to the repository. This change will ensure a standard development environment for all collaborators.